pyproject.toml
[project]
name = "lol-match-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.9.0",
    "ipykernel>=7.2.0",
    "lightgbm>=4.6.0",
    "mlflow>=3.10.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]
